RAC: split into more components
